+++ presentation/src/main/java/com/example/miso/ui/list/component/list/InquiryListItem.kt
@@ -1,6 +1,5 @@
-package com.example.miso.ui.list.component
+package com.example.miso.ui.list.component.list
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
@@ -18,11 +17,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.miso.R
 import com.example.miso.ui.theme.MisoTheme
 
 @Composable
