+++ presentation/src/main/java/com/example/miso/ui/list/component/list/ListTitleText.kt
@@ -1,4 +1,4 @@
-package com.example.miso.ui.list.component
+package com.example.miso.ui.list.component.list
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
